org README: simplify and rely on links
This profile appears at the top of our org page, meaning the longer it is, the more it pushes away access to the code. In the spirit of "less talk, more code," I think we can link out to relevant resources here but make it a lot more concise so folks can get at the code more easily.